requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0.post2
bcrypt==4.1.3
cryptography==41.0.7
pycryptodome==3.23.0
PyMySQL==1.1.1
python-dotenv==0.21.1
gunicorn==21.2.0
